requirements.txt

certifi==2020.6.20
chardet==3.0.4
Django==3.2.3
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
idna==2.10
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
requests==2.24.0
s3transfer==0.2.0
six==1.15.0
urllib3==1.25.11
